user

+++ server/model/user/user_entry_card.go
@@ -26,8 +26,8 @@ type UserEntryCard struct {
 	GiftAmt        uint16     `json:"giftAmt" form:"giftAmt" gorm:"column:gift_amount;not null;default:0;comment:赠送数量;"`
 	SurplusAmt     uint16     `json:"surplusAmt" gorm:"column:surplus_amount;not null;default:0;comment:剩余数量;"`
 	ContractNumber uuid.UUID  `json:"contractNumber" gorm:"column:contract_number;not null;size:191;comment:合同编号;"`
-	TotalPrice     uint       `json:"totalPrice" gorm:"column:total_price;not null;default:0;comment:原价格;"`
-	PayPrice       uint       `json:"payPrice" form:"payPrice" gorm:"column:pay_price;not null;default:0;comment:实际支付价格;"`
+	TotalPrice     uint16     `json:"totalPrice" gorm:"column:total_price;not null;default:0;comment:原价格;"`
+	PayPrice       uint16     `json:"payPrice" form:"payPrice" gorm:"column:pay_price;not null;default:0;comment:实际支付价格;"`
 	ValidTime      uint16     `json:"validTime" gorm:"column:valid_time;not null;default:0;comment:有效次数;"`
 	ValidPeriod    uint16     `json:"validPeriod" gorm:"column:valid_period;not null;default:1;comment:有效期;"`
 	DateUnit       uint8      `json:"dateUnit" gorm:"column:date_unit;not null;default:1;comment:日期单位：1天，2月，3年;"`
